Add files via upload
Added an option for the VM scan to filter which VMs get scanned. Kaiser had a Horizon install that created and deleted hundreds of VM per hour from opening and closing VDI sessions. The filter would allow them to only scan the VMs hosting the VDI images. It uses '-like' in powershell which allows wildcards before and/or after the search text. If the user does not enter anything for searching then it does not filter and it selects all VMs. I could in the future change the '-like' in powershell to '-match' which uses regex instead.